# Improve php_pcov_ignored_opcode() performance by using a bit set