Checking the opcode took significant time in the profile due to all the branching (compiled with GCC 15.1.1).
Instead, use a bit set LUT.

For the following script:
```php
<?php

pcov\start();

for ($i = 0; $i < 1000000; $i++) {
    microtime();// Do something that can't be optimized away
}

pcov\stop();
```

I obtain on an i7-4790:
```
Benchmark 1: php -d extension=$(realpath modules/pcov_old.so) test.php
  Time (mean ± σ):     399.9 ms ±   5.0 ms    [User: 391.1 ms, System: 7.0 ms]
  Range (min … max):   394.1 ms … 410.6 ms    10 runs

Benchmark 2: php -d extension=$(realpath modules/pcov.so) test.php
  Time (mean ± σ):     381.4 ms ±   2.0 ms    [User: 373.6 ms, System: 6.4 ms]
  Range (min … max):   378.4 ms … 384.0 ms    10 runs

Summary
  php -d extension=$(realpath modules/pcov.so) test.php ran
    1.05 ± 0.01 times faster than php -d extension=$(realpath modules/pcov_old.so) test.php
```