@krakjoe 2025-06-29
Thanks @nielsdos, feel free to make improvements ... when happy with the set, open an issue assigned to @remicollet for a release.